+++ eyes.selenium.java/src/test/java/com/applitools/eyes/selenium/TestSetup.java
@@ -103,7 +103,7 @@ public abstract class TestSetup extends ReportingTestSuite implements ITest {
     //protected RectangleSize testedPageSize = new RectangleSize(1200, 800);
     protected RectangleSize testedPageSize = new RectangleSize(700, 460);
 
-    protected boolean compareExpectedRegions = false;
+    protected boolean compareExpectedRegions = true;
 
     protected String platform;
     protected boolean forceFPS;